Move well-known routes into REST module
