+++ src/gateway/rest/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 pub mod a2a;
+pub mod well_known;
